docs: Add typechecking for fastify plugin
