+++ packages/gasket-plugin-fastify/package.json
@@ -13,8 +13,8 @@
     "test": "cross-env NODE_OPTIONS='--unhandled-rejections=strict' jest",
     "test:watch": "jest --watch",
     "test:coverage": "jest --coverage",
-    "posttest": "npm run lint",
-    "typecheck:skip": "tsc",
+    "posttest": "npm run lint && npm run typecheck",
+    "typecheck": "tsc",
     "typecheck:watch": "tsc --watch"
   },
   "repository": {
