+++ packages/gasket-plugin-fastify/tsconfig.json
@@ -16,6 +16,7 @@
   },
   "exclude": [
     "test",
-    "coverage"
+    "coverage",
+    "generator"
   ]
 }
